+++ lib/screens/date_picker_sheet.dart
@@ -6,6 +6,7 @@ import '../widgets/my_container.dart';
 
 /// Date Picker Screen displays option to switch between dates
 /// from 1st of this month to current date of present month.
+@deprecated
 class DatePickerSheet extends StatefulWidget {
   final RouteSettings settings;
   DatePickerSheet(this.settings);
