+++ lib/screens/profile_page/pie_chart.dart
@@ -1,16 +1,9 @@
-import 'package:fl_chart/fl_chart.dart';
-import 'package:flutter/cupertino.dart';
-
-// Files
-import '../controllers/activity.dart';
-import '../models/mood.dart';
-import '../styles/style.dart';
-import '../utils/color.dart';
+part of 'profile_page.dart';
 
 class PieChartWidget extends StatefulWidget {
-  final String type;
+  final AnalyticsOption type;
 
-  PieChartWidget(this.type);
+  PieChartWidget({this.type});
 
   @override
   _PieChartWidgetState createState() => _PieChartWidgetState();
@@ -23,15 +16,18 @@ class _PieChartWidgetState extends State<PieChartWidget> {
       padding: EdgeInsets.symmetric(vertical: 16),
       child: PieChart(
         PieChartData(
-            borderData: FlBorderData(show: false),
-            sectionsSpace: 0,
-            centerSpaceRadius: 0,
-            sections: showingSections()),
+          borderData: FlBorderData(show: false),
+          sectionsSpace: 0,
+          centerSpaceRadius: 0,
+          sections: (widget.type == AnalyticsOption.monthly)
+              ? monthlySections()
+              : weeklySections(),
+        ),
       ),
     );
   }
 
-  List<PieChartSectionData> showingSections() {
+  List<PieChartSectionData> monthlySections() {
     List<PieChartSectionData> chart = [];
     ActivityController.getMonthGraph(2021, 1).forEach(
       (key, value) {
@@ -48,6 +44,24 @@ class _PieChartWidgetState extends State<PieChartWidget> {
         );
       },
     );
+    return chart;
+  }
+
+  List<PieChartSectionData> weeklySections() {
+    List<PieChartSectionData> chart = [];
+    chart.add(
+      PieChartSectionData(
+        color: c.CupertinoDynamicColor.resolve(
+          c.CupertinoColors.secondarySystemBackground,
+          context,
+        ),
+        value: 100,
+        title: 'Coming Soon..',
+        radius: 150,
+        titleStyle: Style.body(context),
+      ),
+    );
+
     return chart;
   }
 }
