+++ lib/utils/enums.dart
@@ -1,0 +1,12 @@
+enum AnalyticsOption {
+  monthly,
+  weekly,
+}
+enum BottomTabOption {
+  home,
+  profile,
+}
+enum ActivityType {
+  add,
+  edit,
+}
